French layout support for cmd-hid
- removed the LANG="C" on start-hid-* scripts to allow dynamic LANG selection
- modified the hid-cmd.php to retreive the browser's default language
Unfortunaly, hid-cmd.php isn't available as a file in the repository,
and is included in a tar.gz archive, so I include both the single file and
the updated tar.gz archive
- modified the keyseed.py to match the right keys for french keyboard
Note: This is my first use of github and git management, so I hope I didn't
break "state of art" rules. Feel free to give my feedback on this point.